+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/psi/impl/xquery/XQueryFunctionDeclPsiImpl.kt
@@ -110,6 +110,7 @@ class XQueryFunctionDeclPsiImpl(node: ASTNode) :
 
     override fun getPresentableText(type: ItemPresentationEx.Type): String? = when (type) {
         ItemPresentationEx.Type.StructureView -> cachedStructurePresentableText.get()
+        ItemPresentationEx.Type.NavBarPopup -> cachedStructurePresentableText.get()
         else -> presentableText
     }